+++ Support/Extensions/Types.cs
@@ -17,6 +17,37 @@ namespace Platform.Support
             return Helpers.IsNull<T>(value, replacement);
         }
 
+        /// <summary>
+        /// Converts an array of bytes to a string Hex representation
+        /// </summary>
+        public static string ToHex(this byte[] ba)
+        {
+            return Helpers.ToHex(ba);
+        }
+
+        /// <summary>
+        /// converts from a string Hex representation to an array of bytes
+        /// </summary>
+        public static byte[] FromHex(this string hexEncoded)
+        {
+            return Helpers.FromHex(hexEncoded);
+        }
+
+        /// <summary>
+        /// converts from a string Base64 representation to an array of bytes
+        /// </summary>
+        public static byte[] FromBase64(this string base64Encoded)
+        {
+            return Helpers.FromBase64(base64Encoded);
+        }
+
+        /// <summary>
+        /// converts from an array of bytes to a string Base64 representation
+        /// </summary>
+        public static string ToBase64(this byte[] b)
+        {
+            return Helpers.ToBase64(b);
+        }
 
     }
 }