+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Console extensions
 - SemanticVersion type
 - OS Detection
+- Encryption Hash Class
 
 ### Fixed
 - Reflection GetVersion bug
@@ -14,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Namespaces simplification
+- CRC32 Algorithm
 
 ## [3.1.9] - [2016-04-25]
 ### Added